requirements.txt
pytubefix >= 9.4.1
requests
python-dotenv
thefuzz
numpy